Move ContractClassGatewaySerializer to separate file
