+++ lib/src/main/kotlin/com/swmansion/starknet/data/types/Contracts.kt
@@ -2,9 +2,6 @@ package com.swmansion.starknet.data.types
 
 import com.swmansion.starknet.extensions.base64Gzipped
 import kotlinx.serialization.*
-import kotlinx.serialization.descriptors.SerialDescriptor
-import kotlinx.serialization.encoding.Decoder
-import kotlinx.serialization.encoding.Encoder
 import kotlinx.serialization.json.*
 
 enum class AbiEntryType {
@@ -139,34 +136,6 @@ data class ContractClass(
         @SerialName("L1_HANDLER")
         val l1Handler: List<ContractEntryPoint>,
     )
-}
-
-internal object ContractClassGatewaySerializer : KSerializer<ContractClass> {
-    @Serializable
-    data class ContractClassGateway(
-        val program: JsonElement,
-
-        @SerialName("entry_points_by_type")
-        val entryPointsByType: ContractClass.EntryPointsByType,
-    )
-
-    override fun deserialize(decoder: Decoder): ContractClass {
-        val response = ContractClassGateway.serializer().deserialize(decoder)
-
-        val programString = response.program.toString()
-        val program = programString.base64Gzipped()
-
-        // FIXME: It doesn't produce the same output as the rpc endpoint
-
-        return ContractClass(program, response.entryPointsByType)
-    }
-
-    override val descriptor: SerialDescriptor
-        get() = ContractClass.serializer().descriptor
-
-    override fun serialize(encoder: Encoder, value: ContractClass) {
-        throw Exception("Class used for deserialization only.")
-    }
 }
 
 @Serializable
